chore: 이미지카드 관련 컴포넌트 폴더 정리
