+++ src/components/modals/photo_exchange/PhotoExchange.jsx
@@ -12,7 +12,7 @@ import { Title } from "../../common/title/Title";
 import SearchBar from "../../common/search_bar/SearchBar";
 import { DropdownNoneBorder } from "../../common/dropdown_normal/DropdownNormal";
 import styles from "./PhotoExchange.module.css";
-import ImgCardMy from "../../imgcard_my/ImgCardMy";
+import ImgCardMy from "../../img_card/imgcard_my/ImgCardMy";
 import PhotoExchangeDetail from "./PhotoExchangeDetail";
 import Filter from "../filter_modal/FilterModal";
 import { useRef } from "react";
